+++ src/document/search.js
@@ -36,6 +36,7 @@ import { highlight_fields } from "./highlight.js";
  * @param {number|DocumentSearchOptions=} limit
  * @param {DocumentSearchOptions=} options
  * @param {Array<Array>=} _promises async recursion
+ * @this Document
  * @returns {
  *   DocumentSearchResults|
  *   EnrichedDocumentSearchResults|
@@ -404,7 +405,7 @@ Document.prototype.search = function(query, limit, options, _promises){
 
             if(count){
                 PROFILER && tick("Document.search:tag:intersect");
-                res = intersect_union(res, arr, resolve); // intersect(arr, limit, offset)
+                res = intersect_union(/** @type {IntermediateSearchResults} */ (res), arr, resolve); // intersect(arr, limit, offset)
                 len = res.length;
                 if(!len && !suggest){
                     // nothing matched
@@ -569,7 +570,7 @@ function merge_fields(fields){
             tmp = group_field[id];
             if(!tmp){
                 entry["field"] = group_field[id] = [key];
-                final.push(/** @type {MergedDocumentSearchEntry} */ (entry));
+                final.push(/** @type {!MergedDocumentSearchEntry} */ (entry));
             }
             else{
                 tmp.push(key);
@@ -618,7 +619,6 @@ function get_tag(tag, key, limit, offset, enrich){
  * @return {EnrichedSearchResults|SearchResults|Promise<EnrichedSearchResults|SearchResults>}
  * @this {Document|Index|WorkerIndex|null}
  */
-
 export function apply_enrich(ids){
 
     if(!SUPPORT_STORE || !this || !this.store) return ids;