+++ src/index/remove.js
@@ -6,6 +6,7 @@ import {
 // <-- COMPILER BLOCK
 import { is_array } from "../common.js";
 import Index, { autoCommit } from "../index.js";
+import { KeystoreMap } from "../keystore.js";
 
 /**
  * @param {!number|string} id
@@ -115,7 +116,7 @@ Index.prototype.remove = function(id, _skip_deletion){
 
 /**
  * When called without passing ID it just will clean up
- * @param {!Map|Array<number|string|Array<number|string>>} map
+ * @param {!Map|KeystoreMap|Array<number|string|Array<number|string>>} map
  * @param {!number|string=} id
  * @return {number}
  */