+++ src/serialize.js
@@ -1,17 +1,22 @@
 // COMPILER BLOCK -->
-import { SUPPORT_STORE, SUPPORT_TAGS, SUPPORT_WORKER } from "./config.js";
+import {
+    SUPPORT_STORE,
+    SUPPORT_TAGS,
+    SUPPORT_WORKER
+} from "./config.js";
+import { IntermediateSearchResults } from "./type.js";
 // <-- COMPILER BLOCK
 import Index from "./index.js";
 import Document from "./document.js";
+import { KeystoreMap, KeystoreSet } from "./keystore.js";
 import { is_string } from "./common.js";
-import { IntermediateSearchResults } from "./type.js";
 
 const chunk_size_reg = 250000;
 const chunk_size_map = 5000;
 const chunk_size_ctx = 1000;
 
 /**
- * @param {Map<IntermediateSearchResults>} map
+ * @param {Map<IntermediateSearchResults>|KeystoreMap<IntermediateSearchResults>} map
  * @param {number=} size
  * @return {Array<Object>}
  */
@@ -34,8 +39,8 @@ function map_to_json(map, size = 0){
 
 /**
  * @param {Array<Object>} json
- * @param {Map<IntermediateSearchResults>} map
- * @return {Map<IntermediateSearchResults>}
+ * @param {Map<IntermediateSearchResults>|KeystoreMap<IntermediateSearchResults>} map
+ * @return {Map<IntermediateSearchResults>|KeystoreMap<IntermediateSearchResults>}
  */
 function json_to_map(json, map){
     map || (map = new Map());
@@ -47,7 +52,7 @@ function json_to_map(json, map){
 }
 
 /**
- * @param {Map<Map<IntermediateSearchResults>>} ctx
+ * @param {Map<Map<IntermediateSearchResults>>|KeystoreMap<KeystoreMap<IntermediateSearchResults>>} ctx
  * @param {number=} size
  * @return {Array<Object>}
  */
@@ -72,8 +77,8 @@ function ctx_to_json(ctx, size = 0){
 
 /**
  * @param {Array<Object>} json
- * @param {Map<Map<IntermediateSearchResults>>} ctx
- * @return {Map<Map<IntermediateSearchResults>>}
+ * @param {Map<Map<IntermediateSearchResults>>|KeystoreMap<KeystoreMap<IntermediateSearchResults>>} ctx
+ * @return {Map<Map<IntermediateSearchResults>>|KeystoreMap<KeystoreMap<IntermediateSearchResults>>}
  */
 function json_to_ctx(json, ctx){
     ctx || (ctx = new Map());
@@ -86,7 +91,12 @@ function json_to_ctx(json, ctx){
 }
 
 /**
- * @param {Set<string|number>|Map<Array<string|number>>} reg
+ * @param {
+ *   Set<string|number>|
+ *   Map<Array<string|number>>|
+ *   KeystoreSet<string|number>|
+ *   KeystoreMap<Array<string|number>>
+ * } reg
  * @return {Array<Array<string|number>>}
  */
 function reg_to_json(reg){
@@ -105,15 +115,23 @@ function reg_to_json(reg){
 
 /**
  * @param {Array<string|number>} json
- * @param {Set<string|number>} reg
- * @return {Set<string|number>}
+ * @param {
+ *   Set<string|number>|
+ *   Map<Array<string|number>>|
+ *   KeystoreSet<string|number>|
+ *   KeystoreMap<Array<string|number>>
+ * } reg
+ * @return {
+ *   Set<string|number>|
+ *   KeystoreSet<string|number>
+ * }
  */
 function json_to_reg(json, reg){
     reg || (reg = new Set());
     for(let i = 0; i < json.length; i++) {
         reg.add(json[i]);
     }
-    return reg;
+    return /** @type {Set<string|number>} */ (reg);
 }
 
 /**
@@ -228,7 +246,7 @@ export function exportIndex(callback, _field, _index_doc = 0, _index_obj = 0){
 
 /**
  * @param {string} key
- * @param {string|Array<Object>} data
+ * @param {string|Array<Object>=} data
  * @this Index
  */
 
@@ -326,7 +344,9 @@ export function exportDocument(callback, _field, _index_doc = 0, _index_obj = 0)
             case SUPPORT_STORE && 2:
 
                 key = "doc";
-                chunk = this.store && map_to_json(this.store);
+                chunk = this.store && map_to_json(
+                    /** @type {Map<IntermediateSearchResults>} */ (this.store)
+                );
                 _field = null;
                 break;
 
@@ -346,7 +366,7 @@ export function exportDocument(callback, _field, _index_doc = 0, _index_obj = 0)
             callback,
             _field,
             key,
-            chunk,
+            /** @type {Array|null} */ (chunk || null),
             _index_doc,
             _index_obj
         );